Added updateUserHandler

+++ Node-BackEnd/controllers/feed.js
@@ -4,6 +4,7 @@ const Post = require(`../models/post`);
 const fs = require(`fs`);
 const path = require(`path`);
 const { body } = require(`express-validator/check`);
+const User = require(`../models/user`);
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
   const perPage = 2;
@@ -14,7 +15,8 @@ exports.getPosts = (req, res, next) => {
       totalItems = Number(count);
       return Post.find()
         .skip((currentPage - 1) * perPage)
-        .limit(perPage);
+        .limit(perPage)
+        .populate("creator", "name");
     })
     .then((posts) => {
       res.status(200).json({
@@ -34,35 +36,42 @@ exports.getPosts = (req, res, next) => {
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error(`Validation failed, entered data is incorrect`);
-
-    throw error;
-  }
-
-  if (!req.file) {
-    const error = new Error(`No image provided`);
+    const error = new Error("Validation failed, entered data is incorrect.");
     error.statusCode = 422;
     throw error;
   }
+  if (!req.file) {
+    const error = new Error("No image provided.");
+    error.statusCode = 422;
+    throw error;
+  }
+  const imageUrl = req.file.path.replace(`\\`, "/");
   const title = req.body.title;
   const content = req.body.content;
-  const imageUrl = req.file.path.replace("\\", "/");
+  let creator;
 
   const post = new Post({
     title: title,
     content: content,
-    creator: {
-      name: `Diyan`,
-    },
     imageUrl: imageUrl,
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
       console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
         post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -71,8 +80,6 @@ exports.createPost = (req, res, next) => {
       }
       next(err);
     });
-
-  // Create post in db
 };
 
 exports.getPost = (req, res, next) => {
@@ -143,23 +150,33 @@ exports.updatePost = (req, res, next) => {
     });
 };
 
-const clearImage = (filePath) => {
-  filePath = path.join(__dirname, "..", filePath);
-  fs.unlink(filePath, (err) => console.log(err));
-};
-
 exports.deletePost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
     .then((post) => {
-      //Check logged in user
+      if (!post) {
+        const error = new Error("Could not find post.");
+        error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      res.status(200).json({
-        message: `Deleted post`,
-      });
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
+      res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -167,4 +184,9 @@ exports.deletePost = (req, res, next) => {
       }
       next(err);
     });
+};
+
+const clearImage = (filePath) => {
+  filePath = path.join(__dirname, "..", filePath);
+  fs.unlink(filePath, (err) => console.log(err));
 };
